Reset the data before execute the http request.
Here is a video fo the current problem,
http://snaps.getmoxied.net/3L3c0n0N1Q1k

As you can see on the transition the data from the previous view is
rendered as if is part of the current state and then after the request
is completed the data is being deleted. I think the correct way is to
cleanup the view on a new transition of state.