Handle erros on the request to 404 URLS
At the moment we can make `n` amount of request to invalid URLS and the
FE will always mantain the current state. Ideally or in a normal
scenario on an invalid request to the view we end up on the 404 page.

With this change we handle the errors created by the $http requests and
move to the 404 view.

Fixes #19